Implement some of the lattice functionality
The Lattice class constructor has been implemented in terms of static
functions that handle both periodic and finite boundary conditions.
An invalid lattice site token was introduced to indicate that a nearest
neighbor is invalid, and will only be checked in the instance that the
boundary condition is finite (as all sites will have full neighbor
arrays in the periodic case).